Use fused MLA RoPE kernels in MultiLatentAttention
Move the Triton MLA RoPE kernels (Megatron-LM
fused_mla_yarn_rope_apply port) from tests/pytorch/attention/
mla_rope_utils.py into models/deepseek_v3/mla_rope.py and use them in
MultiLatentAttention: the q kernel rotates the rope slice in place and
the kv kernel assembles key/value in a single pass, removing the
torch.cat/expand/contiguous copies (~10% of layer GPU time). PyTorch
fallback (same convention) covers missing Triton and bshd.

Fix a latent bug from the test util: the q backward kernel assumed a
contiguous incoming gradient, but cuDNN attention backward can hand
over a strided one (allocator-state dependent IMA). The old test file
stays as a compat shim. Add a Triton-vs-PyTorch parity test.
